Use native vim functions to get char under cursor
A custom function was used to get the character under the cursor. It was
also manipulating registers, which means there is a possibility that it
will overwrite some data in that register, what the user would not
expect.
Instead of the custom function, use builtin vim functions to figure out
the character under the cursor.